# labs/sections/programming-with-python/day-082-a-first-web-api-with-fastapi/expected-output/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bookmarks API",
    "summary": "A small, honest CRUD API used to learn FastAPI on Day 082.",
    "description": "Create, list, read, update and delete bookmarks. Every request body is validated by a pydantic model; every response body is filtered through a response model so internal fields cannot leak.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/bookmarks": {
      "post": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Create a bookmark",
        "description": "201 Created, with a Location header naming the new resource.\n\nThe return annotation is ``StoredBookmark`` \u2014 the object with the secret\nin it \u2014 and that is safe, because ``response_model=BookmarkOut`` filters\nthe response before it is serialized. The handler returns the truth; the\ndeclaration decides what the client is entitled to see.",
        "operationId": "create_bookmark_bookmarks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "List bookmarks, optionally filtered by tag",
        "description": "``tag`` is optional; ``limit`` has a default and a validated range.\n\n``limit`` arrives from the query string as text and is handed to the\nhandler as an ``int``, because the annotation said ``int``. Ask for\n``?limit=0`` or ``?limit=abc`` and the handler is never entered: the\ncaller gets a 422 naming ``limit``.",
        "operationId": "list_bookmarks_bookmarks_get",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only bookmarks carrying this tag.",
              "title": "Tag"
            },
            "description": "Return only bookmarks carrying this tag."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of bookmarks to return.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of bookmarks to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookmarkOut"
                  },
                  "title": "Response List Bookmarks Bookmarks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookmarks/{bookmark_id}": {
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Fetch one bookmark",
        "description": "404 when it is not there \u2014 an answer, not a crash.\n\n``HTTPException`` is how a handler says \"this request has a definite,\nnon-exceptional negative answer\". The client gets a small JSON body with\na ``detail`` string. It does not get a traceback, and it must not: a\ntraceback names your files, your line numbers and your local variables.",
        "operationId": "get_bookmark_bookmarks__bookmark_id__get",
        "parameters": [
          {
            "name": "bookmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bookmark Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "404": {
            "description": "No bookmark with that id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Update part of a bookmark",
        "description": "A partial update: only the fields the caller actually sent change.",
        "operationId": "update_bookmark_bookmarks__bookmark_id__patch",
        "parameters": [
          {
            "name": "bookmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bookmark Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkOut"
                }
              }
            }
          },
          "404": {
            "description": "No bookmark with that id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Delete a bookmark",
        "description": "204 No Content: it worked, and there is deliberately nothing to say.\n\nA 204 body must be empty \u2014 that is what the status code means \u2014 so this\nhandler returns a bare ``Response`` rather than a model.",
        "operationId": "delete_bookmark_bookmarks__bookmark_id__delete",
        "parameters": [
          {
            "name": "bookmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bookmark Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No bookmark with that id"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookmarkCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Title",
            "description": "Human-readable name. Must not be empty."
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "Absolute http or https URL."
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Tags",
            "description": "Up to eight short labels."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "BookmarkCreate",
        "description": "The request body of ``POST /bookmarks``."
      },
      "BookmarkOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "tags",
          "created_at"
        ],
        "title": "BookmarkOut",
        "description": "What a client receives. Deliberately a subset of ``StoredBookmark``."
      },
      "BookmarkUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BookmarkUpdate",
        "description": "The request body of ``PATCH /bookmarks/{bookmark_id}``.\n\nEvery field is optional and defaults to ``None``, which is how a partial\nupdate says \"leave this one alone\". ``model_dump(exclude_unset=True)``\nthen tells you which fields the caller actually sent \u2014 note that this is\ngenuinely different from which fields are ``None``, because a caller may\nlegitimately send a field whose value is null."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}